Add autoboot.bin
When flashed to the second bank of a Rev.2 and
above Skunkboard, whis file will cause the board
to automatically boot whatever is flashed to bank
one. This should work even with 6MB roms
providing the autoboot file is flashed first.

To bypass the autoboot and boot to the Skunkboard
green startup screen again, hold down 'A' on the
controller when turning on or reseting your
Jaguar. To permanently remove the autoboot ROM,
flash a 4MB rom over bank 2, or a smaller rom with
the '-e' option to erase the entire bank.